pyproject.toml
# File: DjangoVerseHub/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "djangoversehub"
version = "1.0.0"
description = "Comprehensive Django learning platform with advanced features"
authors = [{name = "DjangoVerse Team", email = "team@djangoverse.hub"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["django", "web", "framework", "learning", "platform"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "Django>=4.2.0,<5.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "django-filter>=23.0",
    "Pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=4.5.0",
    "celery>=5.3.0",
    "channels>=4.0.0",
    "channels-redis>=4.1.0",
    "django-allauth>=0.54.0",
    "django-debug-toolbar>=4.1.0",
    "django-extensions>=3.2.0",
    "python-decouple>=3.8",
    "whitenoise>=6.5.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.0",
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "django-stubs>=4.2.0",
    "mypy>=1.4.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "coverage>=7.2.0",
]

prod = [
    "gunicorn>=21.2.0",
    "whitenoise>=6.5.0",
    "sentry-sdk>=1.32.0",
    "django-storages>=1.13.0",
    "boto3>=1.28.0",
]

[project.urls]
Homepage = "https://github.com/djangoverse/hub"
Documentation = "https://docs.djangoverse.hub"
Repository = "https://github.com/djangoverse/hub.git"
"Bug Tracker" = "https://github.com/djangoverse/hub/issues"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
skip = ["migrations"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "django_verse_hub.settings.dev"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_verse_hub.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=apps --cov-report=html --cov-report=term-missing"